Change thumbnail handling in PostsService
Changed the return value for when there's no image elements found in the post's content, from returning a default image path to null in the PostsService. This was done to allow flexibility in the handling of posts lacking an image, as it opens up for more specific handling in upper layers of the application.