"Replace th:text with JavaScript for unescaped rendering"
Thymeleaf's th:text was previously used to render the content of the posts, however it did not execute any embedded HTML code. To overcome this limitation and enable the rendering of HTML code if any for better text presentation, the th:text attribute was replaced with JavaScript function 'innerText'. Now, the dynamic text is added directly into the DOM via JavaScript, effectively showing any potential HTML markups in the text content.