Add PostsIndexContent DTO for posts rendering
Created a new PostsIndexContent Data Transfer Object (DTO) to facilitate the process of rendering blog posts on the index page. This DTO will carry necessary post data from the back-end to the front-end, making it possible to directly use Thymeleaf for post rendering and thereby replacing the previous Javascript implementation. This change aims to simplify the code structure and enhance maintainability.