Remove unused import and unnecessary annotation in PostsController
Removed the unused import '@RequestBody' from PostsController. Also removed the unnecessary '@RequestBody' annotation in the 'editPost' method, because the form does not need to be passed in the body of the request but it can be extracted directly from the request parameters. This improves efficiency and cleanliness of the code.