Add Google Analytics to layout.html
Google Analytics script has been added to track and report website traffic. This script will allow us to better understand user behavior on our website and use this data for further improvement of the user experience. The script was added as an asynchronous script, so it won't hinder loading speed of our website.