Reorder meta tags after Google Analytics script in layout.html
The meta tags have been moved after the insertion of the Google Analytics script. This change is done to ensure that the Google Analytics script loads prior to anything else on the page. This optimization helps in more accurate tracking of the user behavior as the script gets loaded even before the page's initial render.