Remove Tagify scripts from head, add Summernote CSS in createPostsForm.html
Removed @yaireo/tagify scripts from the head tags to the bottom of body tag and added Summernote CSS to the head tags in createPostsForm.html. This optimization is performed to decrease the page load time and to ensure that the scripts are loaded synchronously at the end of the body element. Additionally, the Summernote CSS is loaded at the head tag to ensure smoother rendering.