Updated createPostsForm to use full version of Summernote
The createPostsForm.html file was updated to use full version of Summernote instead of the lite version to enhance text editing capabilities in new posts. The order of scripts was slightly adjusted for optimised loading and id was added to the textarea element for better JS targeting. Removed duplicate script tags and adjusted their order for better readability and performance.