Update profile picture and rearrange script imports
Changed the profile picture source and added jquery and bootstrap scripts to the head section of layout.html. These changes were made to ensure scripts run after the full page is loaded and prevent issues if they attempt to interact with not completely loaded HTML elements. The profile image was updated for better user