Add UserPasswordEditRequest DTO
Created a new DTO 'UserPasswordEditRequest' under 'com.blog.web.dto' to handle password update requests by the users. The DTO has fields to hold the user's email, current password, and new password. Inclusion of the original password is for security purposes, to validate that the request indeed originates from the user. A Builder Pattern is used for creating instances of this DTO. This aids in handling the increase in fields in UserPasswordEditRequest in future without the complexity of using constructors.