"Added onclick functions to manage forms"
The 'manageCategory' and 'manageAccount' onclick functions were attached to category and account management controls in both 'manageCategoryForm.html' and 'adminUserSetting.html'. This was implemented to allow direct interaction and improve navigation on the site. The 'adminCategoryManage.js' script was also included to allow manipulation of categories by the admin.
In 'adminUserSetting.html', 'openPasswordModal' function is implemented to show the change password modal form and 'saveChanges'