Update Summernote & Dropdown functionalities in forms
The refactored code aims to simplify the usage of summernote.js, by opting for the 'summernote-lite' version, which removes unnecessary toolbar options, providing a cleaner user interface. The change of class specifications in CSS for the '.note-editor', '.note-dropdown-menu', and '.note-modal-footer' along with '.card-summernote' ensures UI uniformity.
Changes have also been to dropdowns that resolve certain issues that prevent them from functioning properly, with the swap from "data-toggle" attribute to "data-bs-toggle". This updating enhances the overall functionality and responsiveness of the page in editPostForm.html and createPostsForm.html. The changes aim to give users a better experience navigating the webpage.