Adjust CSS for better UI and fix dropdown functionality
Improved the UI by modifying the '.tags' and '.form-control.me-sm-2' class, with changes to margins and border-radius for a more pleasing and streamlined look. Also, the dropdown issue has been resolved by swapping "data-toggle" attribute to "data-bs-toggle". This improves the responsiveness of the page and enhances usability in editing and creating forms.