Adjust layout and styles for posts' presentation
Adjusted the layout and styles in 'post.html' and 'style.css' respectively to improve the presentation of posts. The '.list-group' CSS style was removed as it was causing the posts to stretch wide. The 'post-inner' class was moved to the 'container' div in 'posts.html' to provide more localized styling.