Update meta tags to use plain text content in posts.html
The changes made in the posts.html file replace postFindById.getContent() method calls with a plainTextContent variable. This updates the meta description and og:description tags to use plain text instead of content that may include HTML, in order to enhance the readability and SEO of the web page. This change aims to offer better search engine indexing by providing a cleaner snippet of the content. The length limit for descriptions has been increased from 70 to 100 characters, allowing for more descriptive summaries.