Updated class name in layout for improved responsiveness
The class of the div element containing the 'post-card-head' fragment has been changed from 'test' to 'row'. This change was made in order to match the CSS updates that were implemented recently, which aim at improving the alignment, responsive design, and user experience of the website. These CSS changes also support better coding practices for easier maintenance and modifications in the future.