Update navbar-brand link and top_scroll class name, add Icons8 credit.
Navbar-brand link in the header no longer carries the name 'Grand Aventura', as it was deemed unnecessary repetition that didn't contribute to the intuitiveness of the page. The name of the top_scroll class in the footer section has been edited for better readability and consistency across the CSS file. Additionally, a section has been added to attribute 'Icons8', a third-party icons resource used in the project, in accordance to their usage policy and to ensure correct credit is given.