Update website branding and improve UI animations
This commit removes the "Gran Aventura" brand from the navigation bar and enhances the website's interactivity by adding animations. Added transitions for smoother animations and hover effects to improve UI/UX. Padding properties have been adjusted for an optimized view. Also, the codebase is tidier with the removal of redundant styles that were previously commented out. Lastly, a reference to icons8.com was added in the footer to credit the source of used icons.