Add specific CSS class for Github icon and improve layout
This commit includes changes to both HTML and CSS files. The Github icon in the footer of the template gets a specific CSS class to allow specific styling. Also, changes in the style.css include modifications for better UI - a margin-top was added to the container in the left column, padding-top has been added to the right column, and the size of the github icon has been refreshed. These changes aim to improve the user interface and overall browsing experience on the website.