Add category sidebar to post create and edit forms
In `PostsController.java`, added code to fetch the list of all categories and their associated post creation dates. This information is attached to the model as an attribute, allowing it to be displayed as a sidebar in the "Create Posts" and "Edit Posts" forms. This enhancement supports easier navigation for users when creating or editing posts.