"Refactor category repository and enforce unique list order"
Changed the projection method for creating a CategoryAndPostCreatedDateResponse in CategoryRepositoryImpl from Projections.constructor to a new QCategoryAndPostCreatedDateResponse. Also added an orderBy statement for category.listOrder. This will improve the readability of the code and sort the results in ascending order based on listOrder. Additionally, made listOrder column unique in Category class to prevent duplicate values and these changes increase the efficiency of category retrieval operations.