Add assertions to PostControllerTest and modify category title
The savePost test in the PostControllerTest class is now more robust due to included assertions that confirm the data is stored as expected when a post is created. The category title within the same test was also updated to "Spring2" from "Spring" for more clarity and uniqueness among test data. Methods used for these changes are from the assertJ library, which was also added to the import statements at the start of the file.