Implement custom OAuth2 user service
A new file `CustomOAuth2UserService.java` has been added to handle OAuth2 user authentication. This class loads user information from the OAuth2 User Request and either saves a new user or updates an existing user in the system. It also sets the user's role and other attributes in the session and logging for ease of tracking.